create separate test file for zos stat attribute filter
